Refactor CMakeLists.txt to improve PowerShell detection and update various files to use ssize_t for index and length types. This includes changes in CCommandLine, CNormalProcess, and multiple other files to enhance compatibility and prevent potential overflow issues. Additionally, integrate ssize_compat.h for consistent type usage across the codebase.

+++ sakura_core/_main/CCommandLine.h
@@ -29,6 +29,7 @@
 #include "basis/EditInfo.h"
 #include "basis/GrepInfo.h"
 #include "util/design_template.h"
+#include "util/ssize_compat.h"
 
 /*-----------------------------------------------------------------------
 クラスの宣言
@@ -83,8 +84,8 @@ public:
 	tagSIZE GetWindowSize() const noexcept { return { m_fi.m_nWindowSizeX, m_fi.m_nWindowSizeY }; }
 	tagPOINT GetWindowOrigin() const noexcept { return { m_fi.m_nWindowOriginX, m_fi.m_nWindowOriginY }; }
 	LPCWSTR GetOpenFile() const noexcept { return m_fi.m_szPath; }
-	int GetFileNum(void) const noexcept { return static_cast<int>(m_vFiles.size()); }
-	const WCHAR* GetFileName(int i) const noexcept { return i < GetFileNum() ? m_vFiles[i].c_str() : nullptr; }
+	ssize_t GetFileNum(void) const noexcept { return static_cast<ssize_t>(m_vFiles.size()); }
+	const WCHAR* GetFileName(ssize_t i) const noexcept { return i < GetFileNum() ? m_vFiles[static_cast<size_t>(i)].c_str() : nullptr; }
 	void ClearFile(void) noexcept { m_vFiles.clear(); }
 	LPCWSTR GetDocType() const noexcept { return m_fi.m_szDocType; }
 	ECodeType GetDocCode() const noexcept { return m_fi.m_nCharCode; }
