+++ sakura_core/cmd/COpeBlk.h
@@ -20,6 +20,7 @@ class COpeBlk;
 
 #include "cmd/COpe.h"
 #include <vector>
+#include "util/ssize_compat.h"
 
 /*-----------------------------------------------------------------------
 クラスの宣言
@@ -43,9 +44,9 @@ public:
 	~COpeBlk();
 
 	//インターフェース
-	int GetNum() const{ return (int)m_ppCOpeArr.size(); }	//!< 操作の数を返す
+	ssize_t GetNum() const{ return static_cast<ssize_t>(m_ppCOpeArr.size()); }	//!< 操作の数を返す
 	bool AppendOpe( COpe* pcOpe );							//!< 操作の追加
-	COpe* GetOpe( int nIndex );								//!< 操作を返す
+	COpe* GetOpe( ssize_t nIndex );								//!< 操作を返す
 	void AddRef() { m_refCount++; }	//!< 参照カウンタ増加
 	int Release() { return m_refCount > 0 ? --m_refCount : 0; }	//!< 参照カウンタ減少
 	int GetRefCount() const { return m_refCount; }	//!< 参照カウンタ取得
